refactor: 북마크 상세 리팩토링
